python-crawler/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.10
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
feedparser==6.0.10
python-dateutil==2.8.2
pytz==2023.3
loguru==0.7.2
alembic==1.13.1
beautifulsoup4==4.12.2
openai==1.54.3
jupyter==1.1.1
jupyterlab==4.3.3
ipywidgets==8.1.5
matplotlib==3.9.4
pandas==2.2.3


